requirements.txt
requests
beautifulsoup4
Flask
flask-limiter
python-dotenv